refactor: refactor utils to use named export
