+++ src/lib/instill/connector/source/mutations.ts
@@ -1,4 +1,4 @@
-import { env } from "@/utils/config";
+import { env } from "@/utils";
 import { createInstillAxiosClient } from "../../helper";
 import { Source } from "./types";
