+++ src/lib/instill/model/mutations.ts
@@ -1,5 +1,5 @@
 import { Nullable } from "@/types/general";
-import { env } from "@/utils/config";
+import { env } from "@/utils";
 import { createInstillAxiosClient } from "../helper";
 import { Operation } from "../types";
 import { Model } from "./types";
