+++ src/pages/api/set-user-cookie.ts
@@ -1,5 +1,5 @@
 import { setCookie, SetCookiePayload } from "@/lib/cookie";
-import { env } from "@/utils/config";
+import { env } from "@/utils";
 import { NextApiRequest, NextApiResponse } from "next";
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
