+++ src/pages/pipelines/[id].tsx
@@ -24,7 +24,7 @@ import { useSendAmplitudeData } from "@/hooks";
 import { Pipeline } from "@/lib/instill";
 import { getCodeHikeTemplateSource } from "@/lib/markdown";
 import { MDXRemote, MDXRemoteSerializeResult } from "next-mdx-remote";
-import { env } from "@/utils/config";
+import { env } from "@/utils";
 
 export const getServerSideProps: GetServerSideProps<PipelinePageProps> = async (
   context
