+++ src/utils/index.ts
@@ -1,5 +1,6 @@
-export { getHumanReadableStringFromTime } from "./timeUtils";
 export { getTextWidth } from "./dimensionUtils";
 export { groupBy } from "./arrayUtils";
 export { determineModelState } from "./determineModelState";
 export { env } from "./config";
+
+export * from "./getHumanReadableStringFromTime";
